+++ src/Devices/Humidifier.ts
@@ -341,6 +341,7 @@ export class Humidifier {
     } catch (e) {
       this.platform.log.error(JSON.stringify(e.message));
       this.platform.log.debug('Humidifier %s -', this.accessory.displayName, JSON.stringify(e));
+      this.apiError(e);
     }
   }
 
@@ -380,6 +381,7 @@ export class Humidifier {
     } catch (e) {
       this.platform.log.error(JSON.stringify(e.message));
       this.platform.log.debug('Humidifier %s -', this.accessory.displayName, JSON.stringify(e));
+      this.apiError(e);
     }
   }